FIXME: Warnings

+++ src/common/utils.c
@@ -480,7 +480,7 @@ int utils_create_pid_file(pid_t pid, const char *filepath)
 		goto error;
 	}
 
-	ret = fprintf(fp, "%d\n", pid);
+	ret = fprintf(fp, "%d\n", (int) pid);
 	if (ret < 0) {
 		PERROR("fprintf pid file");
 		goto error;
@@ -489,7 +489,7 @@ int utils_create_pid_file(pid_t pid, const char *filepath)
 	if (fclose(fp)) {
 		PERROR("fclose");
 	}
-	DBG("Pid %d written in file %s", pid, filepath);
+	DBG("Pid %d written in file %s", (int) pid, filepath);
 	ret = 0;
 error:
 	return ret;
